Added playerstate 'LOADING' (#344)
Adds the missing playerstate 'LOADING' that can be set in the
extendedStatus.

I was getting the following exception whenever my chromecast would send
a status message with that state:
```
Newtonsoft.Json.JsonSerializationException: Error setting value to 'PlayerState' on 'Sharpcaster.Models.Media.MediaStatus'. ---> System.NullReferenceException: Object reference not set to an instance of an object.
   at SetPlayerState(Object , Object )
   at Newtonsoft.Json.Serialization.DynamicValueProvider.SetValue(Object target, Object value)
```

Docs:
https://developers.google.com/cast/docs/reference/web_receiver/cast.framework.messages.MediaStatus#extendedStatus

Co-authored-by: Grigly <grigly0@gmail.com>