Add configurable playlists to console app (#384)
My first take as basis for discussions.

I added micrososfts configuration.json package to parse
application.json. Hope this is compatible with the AOT and multi target
goal.

Not yet clear how to really structure the config. I tend to have a
List<playlists> to be able to have one level of 'category' like
"Classic-Cds", "Pop-Cds", "Web-Radios".

On the other hand I want to avoid to much keystrokes to navigate. So
That would be a Command line switch where I need to address somehow
which playlist to use. (by "cat"-"Name" or some - self generated unique
IDs?). Or I could start with a category only and directly reach the
according playlist selection menu....